Forced Display Modes moved to Cores Service
- added cores service to global helper
- moved event args models into their own files
- renamed analogue firmware service to firmware service
- renamed image packs service to platform image packs service
