+++ src/models/Base.cs
@@ -44,11 +44,3 @@ public class Base
         handler?.Invoke(this, e);
     }
 }
-
-public class StatusUpdatedEventArgs : EventArgs
-{
-    /// <summary>
-    /// Contains the message from the updater
-    /// </summary>
-    public string Message { get; set; }
-}
