+++ src/models/StatusUpdatedEventArgs.cs
@@ -1,0 +1,9 @@
+namespace Pannella.Models;
+
+public class StatusUpdatedEventArgs : EventArgs
+{
+    /// <summary>
+    /// Contains the message from the updater
+    /// </summary>
+    public string Message { get; set; }
+}
