+++ src/services/FirmwareService.cs
@@ -6,7 +6,7 @@ using Pannella.Models.Analogue;
 namespace Pannella.Services;
 
 [UnconditionalSuppressMessage("Trimming", "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code", Justification = "<Pending>")]
-public class AnalogueFirmwareService : BaseService
+public class FirmwareService : BaseService
 {
     private const string BASE_URL = "https://www.analogue.co/";
     private const string DETAILS = "support/pocket/firmware/{0}/details";
