+++ src/services/PlatformImagePacksService.cs
@@ -9,7 +9,7 @@ using File = System.IO.File;
 namespace Pannella.Services;
 
 [UnconditionalSuppressMessage("Trimming", "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code", Justification = "<Pending>")]
-public class ImagePacksService : BaseService
+public class PlatformImagePacksService : BaseService
 {
     private const string END_POINT = "https://raw.githubusercontent.com/mattpannella/pupdate/main/image_packs.json";
